Pick up recent updates from "public" submodule
Generally this would only be updated via changes to this project,
but the CNAME required a change to the submodule itself.

To do this, simply navigated to the submodule directory and did a
"git checkout main". What it had checked out initially was the
previous commit. This basically shows that submodules keep track
of the commit.